use AbstractComponentIT, rename test namespace to tests
