+++ vaadin-custom-field-flow-parent/vaadin-custom-field-flow-integration-tests/src/test/java/com/vaadin/flow/component/customfield/tests/ErrorIT.java
@@ -1,17 +1,18 @@
-package com.vaadin.flow.component.customfield.test;
-
-import com.vaadin.tests.AbstractParallelTest;
+package com.vaadin.flow.component.customfield.tests;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.vaadin.tests.AbstractComponentIT;
 import com.vaadin.flow.component.customfield.testbench.CustomFieldElement;
+import com.vaadin.flow.testutil.TestPath;
 
-public class ErrorIT extends AbstractParallelTest {
+@TestPath("vaadin-custom-field/error")
+public class ErrorIT extends AbstractComponentIT {
     @Before
     public void init() {
-        getDriver().get(getBaseURL() + "/vaadin-custom-field/error");
+        open();
     }
 
     @Test
